auto_agent/requirements.txt
python-dotenv
openai
google-api-python-client
google-auth
google-auth-oauthlib
requests
fastapi
uvicorn
